FixMobileTopMarginR6.js
6th Attempt:
Changed class query to id query

Table Reformatting at Media Query was intended for tables in the main section, but was mistakenly allowed to apply to tables within modals; this resulted keeping the content too wide on mobile.  This commit revises the CSS to exclude reformatting tables that are within the modals.