FixMobileTopMargin_R8.js
8th Attempt:
This attempt creates a --vh variable for styling to be used in place of the built-in vh.  the built-in vh on miboile devices is now thought to be the issue; 100vh represents the height of the entire screen, not the window.

Table Reformatting at Media Query was intended for tables in the main section, but was mistakenly allowed to apply to tables within modals; this resulted keeping the content too wide on mobile.  This commit revises the CSS to exclude reformatting tables that are within the modals.