FixGenericEmailNameColumn.js
Generic Email Table Name column was named 'Name" in the HTML.  This Did not match the JS functions and was causing problems when  creating a new email: the new email name was blank.  Problem has been fixed by changing the column name to "Email_Name" in the initialization of the column_object.